% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdisc.R
\name{canonical}
\alias{canonical}
\title{Canonical representation of a \code{data.frame}}
\usage{
canonical(x, ..., uid = NULL, domain = NULL)
}
\arguments{
\item{x}{A \code{data.frame}.}

\item{...}{Ignored.}

\item{uid}{The unique \code{usubjid} to keep.}

\item{domain}{The domain name, which will be stripped from column name if
present as a prefix.}
}
\value{
A \code{data.frame}.
}
\description{
The canonical representation of a \code{data.frame} makes it easier to merge data
from different SDTM domains. Specifically, there is a CDISC convention that
dictates the many fields in SDTM data sets have the domain abbreviation as
column name prefix (e.g. lbtest, lbdtc, etc). If you want to merge these
data sets using \code{rbind} for example, it is easier if those prefixes are
stripped. This function also harmonizes the \code{usubjid} in different data sets
to a specific subset. Lastly, "empty" columns (containing only \code{NA}) are
stripped (see \link{prune}).
}
