% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdisc.R
\name{isSF}
\alias{isSF}
\alias{isSCREENFAILURE}
\title{Check for the value "SCREEN FAILURE", ignoring \code{NA}}
\usage{
isSF(x)

isSCREENFAILURE(x)
}
\arguments{
\item{x}{A vector of type \code{character} or \code{factor}.}
}
\value{
A vector of type \code{logical}.
}
\description{
Check for the value "SCREEN FAILURE", ignoring \code{NA}
}
