% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdisc.R
\name{partial_date}
\alias{partial_date}
\title{Handle partial dates}
\usage{
partial_date(x, impute = c("missing", "first", "mid", "last"))
}
\arguments{
\item{x}{A character vector containing a representation of a vector of dates
in \code{yyyy-mm-dd} format, some of which may be incomplete.}

\item{impute}{Whether and how to impute the missing information (see Details).}
}
\value{
A \link{Date} object.
}
\description{
In CDISC datasets, dates can sometimes be incomplete due to missing
information.
}
\examples{
dates <- c("2022-03-21", "2022-03", "2022-03-UNK", "2022", "2022-UNK-UNK")
partial_date(dates)
partial_date(dates, impute="first")
partial_date(dates, impute="mid")
partial_date(dates, impute="last")

dates <- sprintf("2024-\%02d", 1:12)  # Handles leap years correctly
partial_date(dates, impute="last")
}
