% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdisc.R
\name{prune}
\alias{prune}
\title{Prune away columns that don't contain any non-missing values}
\usage{
prune(x)
}
\arguments{
\item{x}{A \code{data.frame}.}
}
\value{
A \code{data.frame}.
}
\description{
Prune away columns that don't contain any non-missing values
}
\examples{
x <- data.frame(a=1:10, b=NA, c=c(NA, 1:9))
prune(x)
}
