% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdisc.R
\name{qw}
\alias{qw}
\title{Character vector without quotation marks}
\usage{
qw(...)
}
\arguments{
\item{...}{Any number of arguments, that will be returned as strings (i.e., quoted).}
}
\value{
A \code{character} vector.
}
\description{
Inspired by Perl's \verb{qw/string/}. Also similar to the \code{Cs()} and \code{.q()}
functions from the \code{Hmisc} package.
}
\examples{
qw(a, b, c)
}
